requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
discord.py==1.5.1
pymongo==3.1
requests==2.25.1
utils==1.0.1
websocket-client==0.57.0